Implement fixed sort in tasks reducer
